pfsearch,Parser: Only define `minCost` on nodes
Refactors `calcHeuristicCosts` to only define `minCost` on nodes.
Previously, `minCost` was defined on nodes as well as subnodes, which
had the benefit of subnodes for binary rules containing the sum of the
minimum cost of both of its subnodes as its minimum cost. This change
reduces the number of property definition in `calcHeuristicCosts`,
reducing its portion of processing from 3.3% → 3.1%.

Removes the definition of `minCost` on subnodes in `Parser`, however,
this not change performance because defining a property as `undefined`
surprisingly does not impact performance.

Amends `pfsearch` to add to a path's minimum cost from its next node
(and the second next node if binary), instead of just checking the
passed subnode. While this adds operations, it removes the addition of
`sub.minCost` for terminal nodes, which was always 0.

This change does not impact performance because it is small.

Updates the order of expected semantics, which changed arbitrarily due
to parse trees of identical cost that are reorded due to artifacts of
JavaScript's floating point precision limit.