edit-rules: Add `genGramProps()`
Adds `genGramProps()` to `createEditRules`, which abstracts the
operation previously in `createInsertionRules()` that creates a new
grammatical properties object for an insertion rule containing only the
original rule's grammatical properties applicable to the insertion
rule's non-inserted symbol. All other properties apply only to the
inserted symbol, and are used in its conjugation when later defining
the insertion text. This is ensured because new rules are verified to
only include grammatical properties if the RHS produces accepting
terminal rules.