edit-rules: Check for unreachable symbols
Extends `removeUnusedNonterminalSymbols` with a check and removal of
nonterminal symbols unreachable via the grammar's start symbol. This
check is necessary because it is possible the removal of a rule-less
symbol forces the removal of a binary rule whose other RHS symbol was
that symbol's only path to the start symbol.