edit-rules: Refactor `removeUnusedNonterminalSyms`
Refactors `removeUnreachableNonterminalSyms()` in
`removeUnusedNonterminalSyms` by accumulating unreachable symbols and
only iterating through the grammar once instead of deleting `symbols
immediately. Otherwise, some symbols might be removed here, and alerted
in the console as unreachable, as a result of removing another symbol
here, though the mistake needing correction originates with the initial
symbol. Though all removals are due to grammar design error, and the
latter case is only avoidable on the first invocation of this function,
this procedure emphasizes the needed corrections by reducing console
warnings for symbols removed as a consequence of the problematic
symbols.