edit-rules: Remove ill rules in `semanticChecks`
Extends `semanticChecks.ruleMissingReducedSemantic()` to remove the
provided rule from the grammar if it a non-edit rule (violating
edit-rules are not added before invoking the method).

Extends `getAncestorWithLHSSemantic()` to if each parent rule has been
evaluated for producing any required semantic. If not, recursively
evaluates that rule first and removes it from the grammar if necessary
(i.e., if can not produce a required reduced semantic). Must check
parent rules first in order to only remove the problematic rules that
always fail, and not the descendant rules that only fail sometimes
because of the problematic parent rule.

Extends `semanticChecks.ruleMissingReducedSemantic()` adds
`semanticSafe` property to rules it has checked and not removed.

Adds `semanticChecks.deleteSemanticChecksProps()`, which deletes
instances of the (temporary) rule property, `semanticSafe.

Extends `createEditRules` to invoke
semanticChecks.deleteSemanticChecksProps()` after adding edit-rules,
which relies on the `semanticSafe` properties to avoid edit-rules that
can not produce a required reduced semantic.

Refactors `checkForMissingSemantics()` in `createEditRules` to not
throw an exception for rules
`semanticChecks.ruleMissingReducedSemantic()` reports to not produce
required semantics, because that method recursively searches and
removes the violating rules while avoid descendant rules that only fail
sometimes because of the problematic parent rule.

Extends `checkForMissingSemantics()` to invoke
`removeUnusedNonterminalSyms()` after
`semanticChecks.ruleMissingReducedSemantic()` to remove any rule-less
or unreachable nonterminal symbols that result from the removal of
semantically problematic rules.

Given this procedure, if a `Category` is created, which adds the
essential base rules, but has neither entities nor rules specific to
the category, then all rules and symbols for that category are removed
because none can produce valid parse trees.