grammar: Extend `getAncestorsWithLHSSemantic()`
Extends `getAncestorsWithLHSSemantic()` in `semanticChecks` to check if
no parent rules are found at the end of the function. This function
either determines all parent rules require the base rule to produce a
reduce semantic, or returns `false` before this check if at least one
parent rule does not require such, demonstrating the base rule is
semantically legal in that path.