grammar: Add `appendAncestorPaths()`
Adds `appendAncestorPaths()` to `semanticChecks`, which abstracts the
operation previously in `getAncestorsWithLHSSemantic()` that appends a
parse tree path node for the provided nonterminal symbol and its rule
to each parse tree path in the provided ancestor paths. For
constructing graph representations of the ancestor paths that require
the provided rule produce a reduced semantic.
`getAncestorsWithLHSSemantic()` invokes the function after a recursive
call to a parent rule returns the parse tee paths.