edit-rules: Amend handling of ambiguous edit-rules
Amends `rulesAreAmbiguous()` in `editRulesUtil` to only throw an
exception, regardless of the `stopAmbiguity` argument, when an
edit-rule is ambiguous with a non-edit rule and the edit-rule is
cheaper. Previously, the function threw an exception regardless of the
cost difference.

This enables the function to stop the following insertion (first line),
which is ambiguous with the non-edit rule (second line).
- [repository-plural] -> [repository-no-relative]
([repository-relative-clause] -> "that are repos")
- [repository-plural] -> [repository-no-relative]