edit-rules: Prevent meaningless clause insertions
Re-implements prevention of insertions that produce meaningless clauses.

Reduces previous `isUselessInsertion()` implementation to a single
check, and invokes `isUselessInsertion()` from within
`mergeInsertionSets()`. The latter prevents the construction of
previously (incorrectly) undetected insertions.

Extends `mergeInsertionSets()` with parameters necessary for
`isUselessInsertion()` checks.

Adds extensive (700-word) documentation that includes formal
grammatical and linguistic explanations of the logic of the meaningless
clause insertions and their detection.

It was enormously difficult to solve and formalize this. The
explanation is necessary as its articulation proves the logic's
accuracy, which grammar implementations (i.e., tests) can not as
confidently and rigorously confirm. This final description required 4
hours to prove. The challenge was more about linguistic theory and
grammar than algorithms, and is certainly not computer science.