NSymbol: Get existing `NSymbol` for binary rules
Extends `NSymbol.newBinaryRule()` to check and return if an `NSymbol`
with a single binary rule identical to provided rule definition options
already exists, otherwise create a new `NSymbol` and rule as normal.

Adds `getExistingBinaryRuleLHSSym()`, which gets the existing
`NSymbol`, if any, for which the following are true:
1. Its name is identical to the name generated from the provided
`NSymbol` name tokens.
 * 2. Its only rule is binary and has properties identical to the rule
generated from the provided rule definition options.