NSymbol: Add `grammar.newVerb()`
Adds the module `terminalRuleSetMethods`, which contains the method
`grammar` method `grammar.newVerb()`, which creates an `NSymbol` that
produces terminal rule sets for a verb with the necessary text forms
for conjugation.

This new methods generates the same rules as
`NSymbol.prototype.addVerb()`, but is a much better designed API,
including parametrization of all verb forms of substitutions.

The module's methods which create an `NSymbol` are preferable to
`NSymbol` instance methods that add the same terminal rule sets to an
existing `NSymbol`. By not exposing the `NSymbol` (as easily), this
implementation seeks to avoid mixing these sets' rules with others on
the same symbol.

Extends `grammar` with the assignment of the `terminalRuleSetMethods`
methods to the module.