NSymbol: Add `grammar.newNoun()`
Adds `grammar.newNoun()`, which creates an `NSymbol` that produces
terminal rule sets for a noun with display text for the correct
inflections.

Each terminal rule in each set has the plural noun form of the same set
as its `text` string. When constructing parse trees, `pfsearch` uses
this `text` value as the display text for both rules in this set.

The method's parameter definition groups the terminal rules into sets
for each noun, with each noun form defined for each set, to enable
substitution of incorrect forms in input with the correct form of the
same noun.

Extends `grammar` documentation for the new method.