Category: Extend `prototype.addVerbRuleSet()`
Extends `Category.prototype.addVerbRuleSet()` and
`Category.prototype.addSubjectVerbRuleSet()` with the
`options.noPresentPerfectNegative`, which specifies omitting omitting
the present-perfect-negative rules from the verb rule set. For use by
verbs that represent past actions that can not reoccur for the same
objects, which these rules would otherwise suggest.

If `options.noPresentPerfectNegative` is `true` for
`Category.prototype.addVerbRuleSet()`, which also calls
`Category.prototype.addSubjectVerbRuleSet()` with the same parameter
value, it omits the following rules:
1. (repos) `[nom-users+]` have/has not created
2. (people who) have not created `[repositories+]`

For rule #1, `repositories-created()` returns a list of repositories
which the specified users did not create; however, the rule's
present-perfect tense suggests the specified users can create those
same repositories in the future. This implication is false because a
repository's creation can not reoccur.

For rule #2, `repository-creators()` returns a list of users that did
not create the specified repositories; however, the rule's
present-perfect tense suggests those same users can create the
specified repositories in the future. This implication is false because
a repository's creation can not reoccur.

Though, for rule #2, it *is* possible for the same users which the
query returns to meet the same specified criteria in the future (e.g.,
"repos `{user}` likes"), it will not be for the same objects (i.e.,
"`{user}`" must "like" different repositories in the future). Hence,
this rule's implication is not absolutely false, but it is best to
avoid the uncertainty completely because the same query is possible via
the simple-present-negative rule (e.g., "(people who) did not create
`[repositories+]`").