semantic: `reduceUnion()`
Adds `reduceUnion()`, which abstracts the operation previously in
`semantic.reduce()` to reduce a LHS semantic with a `union()` semantic
node by distributing the LHS semantic function among the `union()`
semantic arguments, creating a new semantic tree with `union()` as the
root.

Refactors `semantic.reduce()` to integrate `reduceUnion()`, and updates
documentation.