rules: Remove one root `union()` rule sets
Removes the rules for conjunction rule sets that used a single
`union()` semantic for conjunction subtrees that used one or more
instances of `[or]`, restoring the previous rule design.

This grammar design, as well as the derivate designs for
`[nom-users-plural-subj]` in `user` and `[cat-filter+]` in `Category`,
provided no benefit.

First, semantic reduction still requires to flatten instances of
`union()` to detect duplicate semantic arguments. For example, the
first query below uses two conjunction subtrees, and will need
flattening of nested `union()` even with this rule design to detect its
semantic equivalence to the second query:
  "my repos I or Danny like or my followers like"
  "my repos I or Danny or my followers like"

Second, the use of these rule increases `Parser.prototype.addNode()`
invocations for the `union` tests in the test suite by 5.5%, and the
total benchmark time for the `union` tests increases by 5% (i.e., 5%
slower).

Third, the hypothesis that knowing every `union()` instance is the root
of a conjunction subtree does not aid in determining how to properly
distribute LHS semantics that among the `union()` semantic children.

Archives the three implementations of this design to
`/lib/old/conjunctionsOneRootUnion/`.