semantic: Fix `union()` reduction with nested LHS
Moves the recursive invocation of `semantic.reduce()`, which reduces a
LHS semantic that has children, to after `reduceUnion()`. This properly
moves a `union()` semantic up a semantic tree and distributes the
entire `lhs` among the `union()` arguments. For example:
  "repos Danny or Aang do not like"
  LHS: `not(repos-liked())`
  RHS: `union(0,1)`
  → union(not(repos-liked(0)),not(repos-liked(1)))`

Updates the documentation accordingly.

Updates the expected values of existing tests to reflect this change.