rules: Add `workAt` for the `company` category
Adds `workAt`, which contains the beginnings of the rule set for
`[work]` in the `company` category. The action which `[work]`
represents is distinguishable from verbs/actions seen thus far because
the tense/temporality of the action is material. This introduces the
first tense semantic, `ever-past()`.

Adds tests and updates the expected values of existing tests for these
new rules.