rules: Temporarily change `[work]` insertion cost
Temporarily increases `[work]` insertion cost to prevent its use with
the `[contribute-to]` substitution set, "work on", which otherwise
yields a lower insertion cost than currently specified for
`[contribute-to]`. Will be addressed when redesigning the splitting of
multi-token terminal rules.

Updates the expected values of existing tests to reflect this change.