Parser: Maintain input tense following deletions
Fixes `Parser.prototype.addTermRuleNodes()` in `matchTerminalRules` to
maintain terminal symbol input tense following a deletion. For use by
parent rules with `acceptedTense`, which specifies a tense to accept
when input but not to enforce if otherwise (i.e., tense is optional
because semantically meaningless).

For example: "repos I do liked" → "repos I liked".