edit-rules: Tweak insertions for substitutions
Tweaks `createInsertionRules` style for creating insertion rules for
nonterminal, multi-token substitutions. For these rules, `pfsearch`
uses the rule's `text` instead of the `text` of the matched terminal
rules it produces. For example:
   `[contribute-to]` → "work" (input) → `[work]` `[on]`, text: `[
{contribute-verb-forms}, "to" ]`

`calcHeuristicCosts` will, however, traverse this rule's child nodes to
get the input tense of any matched verb terminal rules to maintain
tense for any verb substitutions whose parent rule has matching
`acceptedTense` (i.e., maintain the tense if input, but do not
conjugate to the tense if otherwise). For example:
  `[contribute-to]` → "worked" (input) → `[work]` `[on]`, text: `[
"contributed to" "to" ]`