pfsearch: Support binary term set insertions
Creates a new `ruleProps` for the provided binary term set insertion
subnode with the following properties:
1. terminal - The new `ruleProps` is marked terminal so that `pfsearch`
uses this subnode's new `ruleProps` to generate display text and does
not traverse its child nodes.
2. cost - Uses `cost`, the cumulative cost (including any deletion
costs) of the subtree this subnode produces, which will be inaccessible
after assigning this new, terminal `ruleProps`
3. semantic - Keeps this subnode's `ruleProps.semantic`, if any.
4. text - Merges `text` values of the matched terminal rules this
subnode's single child node produces with this subnode's
`ruleProps.text` according to this subnode's `ruleProps.insertedSymIdx`.
5. tense - Gets the tense of any matched verb terminal rules this
subnode produces to maintain tense if the parent rule of this subnode
has matching `acceptedTense`.

For example:
  `[contribute-to]`
    → `[contribute]`, text: "to"
    → "contribute" (input), text: `{contribute-verb-forms}`
    → text: `[ {contribute-verb-forms}, "to" ]` (merged text values)