NSymbol: Save `defaultText` in `g.newVerbSet2()`
Extends `terminalRuleSetMethods2.newVerbSet2()` to assign the term set
sequence's `defaultText` to the returned `NSymbol`. For use when nesting
the `NSymbol` in another term set sequence, which uses this
`defaultText` as part of its default text (if in the first accepted
rule), which is the display text for the new sequence's its substitution
rules.