NSymbol: Detect nested verb term sequences
Extends `terminalRuleSetMethods2.newBinaryTermSequence()` with the
parameter `_isNestedTermSequence`, which specifies `options.termPair`
was defined as a nested term pair within another term sequence pair
passed to this method or `terminalRuleSetMethods.newTermSequence()`. If
`true`, manually checks if `options.termPair` contains a verb to
determine the `isVerb` value for the returned `NSymbol`.
`options.isVerb` exists only to check if an entire sequence contains or
lacks a single verb, not every bigram within the sequence; e.g., "in
common" within the verb-phrase "have in common" lacks a verb.