NSymbol: Validate nested term sequence pairs
Extends `flattenTermSequencePair()` to not throw an exception if a
nested term pair within verb sequence does not contain a verb. This is
necessary `isVerb` only specifies if the entire base sequence contains
or lacks a single verb, not every bigram within the sequence. For
example, this function does not check if "in common" within the verb
sequence "have in common" contains verb; ergo, `isVerb` is unavailable
should the nested pair be "have in", though the pair contains a verb.

Removes `isIllFormedVerbTermSequencePair()` because the checks are not
moved to `flattenTermSequencePair()` to use check in accordance with
whether the provided term pair was defined as nested within another
term sequence.