NSymbol: Fix `isFutileGramProp` for term sequences
Extends `NSymbol.isFutileGramProp()` to traverse the RHS symbols of
term sequences that lack text (i.e., neither a nonterminal substitution
nor an edit-rule) to determine if the provided grammatical property
will be used. `calcHeuristicCosts` merges the `text` values such rules'
RHS symbols produce and assigns it to the rule for `pfsearch` to
conjugate.