Parser: Use absolute value of negative integers
Extends `parseIntToken()` in `matchTerminalRules` to use the absolute
value of input integers.

For now, the grammar forbids range-based integer symbols to have
negative lower bounds because it is unlikely any will be defined as
such. Otherwise, negative input integers will likely fail to match any
integers (because none defined), leading to a re-parse with awful
suggestions.