rules: Refactor `negation` with term sequences
Refactors `negation` rule sets to use term sequences. Will be improved
further, however, to properly implement stop-words.

This change reduces the number of paths created for many tests (because
"not" is a common term). Updates order of expected semantics of these
tests, which changed arbitrarily due to reordering of parse trees of
identical cost.