Category: Restructure `[do-negation]` rule
Replaces a use of `[do-past-negation]` with `[do-negation]` and moves
the necessary `grammaticalForm` for past tense. Possible following the
`gramProps` redesign.

Removes many instances of ambiguity caused by the deletable "do" as a
deletable and its reinsertion, because `flattenTermSequence` flattens
such multiple paths.

Updates the expected values of existing tests to reflect this change.