rules: Refactor `[do-negation]`
Removes the binary rule specifically for "did not" and uses the rule
for "`[do]` not" in its place, with the necessary `grammaticalForm`
value for 'past'.

This grammar structure change yields more parse-time conjugation (i.e.,
worse performance). With the other structure, `[do]` is conjugated to
past tense when flattening, which does not occur with this structure
because the conjugating rule is not a term sequence. This structure's
only benefit is it reduces the number of parser reductions. It is
difficult to compare the performance of the two implementations.