Category: Restructure "did not" rules
Restructures the rule in `Category.prototype.addObjectVerbRuleSet()`
for "did not", by pairing `[do-negation]` with the verb, instead of
with `[nom-users+]`, yielding more rules but far more term sequence
flattening, which reduces the number of paths `pfsearch` creates and
the number of conjugation operations.

Updates the expected values of existing tests to reflect this change.