util: Add `util.skipFileInLocationRetrieval()`
Adds `util.skipFileInLocationRetrieval()`, which marks the module in
which the method is invoked for `util.getModuleCallerLocation()` to
skip when searching the call stack. Extends
`util.getModuleCallerLocation()` to support this method.

This is useful for using `util.getModuleCallerLocation()` to include a
method's invocation location in an error message, though that location
is deep within the call stack relative to the error's generation
location. I.e., the error is caught several modules deep from the
invocation to which the error applies, as opposed to being one module
deep.