term: Add `addVerbTerminalRuleSet()`
Extends `verbTermSet` with the addition of `addVerbTerminalRuleSet`,
which adds a verb terminal rule set to the provided `NSymbol` for every
terminal symbol in the provided verb forms term set, with the provided
display text as each rule's `text` value.

Extends `verbTermSet.newVerb()` and `verbTermSet.newTenseVerb()` to
invoke the function with the necessary values.