rules: Add `conjunction.createDisjunctionSet()`
Adds `conjunction.createDisjunctionSet()`, which creates a logical
disjunction set that produces recursive rules for multiple instances of
the provided symbol separated (only) by "or" (`union()`).

For use by rule sets whose associated semantic forbids multiple
instances of that semantic within an `intersect()` (e.g.,
`repositories-created()`), and therefore forbids logical conjunctions.
That semantic is to use the semantics this new disjunction rule set
produces as its semantic arguments (except for `union()`, which is
moved up the semantic tree to distribute the outer LHS semantic among
the `union()` semantic arguments).

Adds term sequence `[or-substituted-and]`, which substitutes "and" →
"or, for use with disjunction rule sets.