rules: Add `[of-poss-users+-no-intersect]`
Defines `[of-poss-users+-no-intersect]`, the disjunction set for
possession.

 For use by categories who semantic representing possession of
instances of that category's associated DB object is defined with
`forbidsMultiple`. Prevents the intersection of multiple instances of
this ownership semantic because it implies the associated DB object can
be owned/authored by multiple users. Only enables the logical
disjunction of multiple instances of said semantic.