rules: Fix problematic shared rules in `poss`
Amends the sharing of `NSymbol` rules in `poss` to clone the rule
objects to prevent mutation of a shared object. Specifically, the
previous implementation ran `NSymbol.diversifyRuleCosts()` on the same
rules twice.

Updates order of expected semantics of existing tests, which changed
arbitrarily due to reordering of parse trees of identical cost.