semantic: Refactor conjunction semantic completion
Adds `completeSingleNodeConjunctionSemantic()`, which marks instances
of `intersect()` or `union()` in a single node semantic array as
complete. Replace `markRHSConjunctionComplete()` invocation in
`semantic.reduce()` with the new function, optimized to expect a
single-node array.

Adds `completeMultiNodeConjunctionSemantic()`, which marks instances of
`intersect()` complete in the provided multi-node semantic array. For
use by `completeSingleNodeConjunctionSemantic()` with the children of a
`union()` semantic when attempting to complete an already complete
`union()`.