rules: Refactor `repository` to use `headNoun`
Refactors the instantiation of the `repository` `Category` with
`headNoun` instead of `headTerm`, defined as a term sequence of type
'noun'. Updates `repository.headNoun` references accordingly.

Adds a test for the new count nouns with singular forms for the
`repository` category.

Updates the expected values of existing tests to reflect this change.