rules: Refactor `company` to use `headNoun`
Refactors the instantiation of the `company` `Category` with `headNoun`
instead of `headTerm`, defined as a term sequence of type 'noun'.
Updates `company.headNoun` references accordingly.

Adds a test for the new count nouns with singular forms for the
`company` category. Updates the expected values of existing tests to
reflect this change.