NSymbol: Add `termSequenceRHSIndexes` property
Extends `assignTermSequence()` in `NSymbol` to assign
`termSequenceRHSIndexes` to nonterminal rules, which is the RHS indexes
of which RHS symbols are term sequences, irrespective of
`isTermSequence`. For use by `createInsertionRules` to mark insertion
rules as term sequences for which the original rule is not because the
inserted RHS symbol is not a term sequence. `removeTempRulesAndProps`
removes the property upon completion of grammar generation.

Updates documentation accordingly.